pyproject.toml
[project]
name = "fin_analysis_loader"
version = "0.1.0"
description = "Bulk data loader for A-share 1-minute data"
requires-python = ">=3.10"
dependencies = [
    "agent-framework==1.0.0b260130",
    "dashscope>=1.20.0",
    "httpx[socks]>=0.28.1",
    "mplfinance>=0.12.7",
    "numpy>=1.24.0",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.0.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.0",
    "scipy>=1.10.0",
    "socksio",
    "tavily-python>=0.7.21",
    "tqdm>=4.66",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-mock>=3.15.0",
]
